% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survsmart-package.R
\docType{package}
\name{survsmart-package}
\alias{survsmart}
\alias{survsmart-package}
\title{survsmart: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ubesp-dctv.github.io/survsmart/}
}

}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@ubep.unipd.it} (\href{https://orcid.org/0000-0002-0520-7428}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Giulia Lorenzoni \email{giulia.lorenzoni@unipd.it} (\href{https://orcid.org/0000-0003-1771-4686}{ORCID}) [copyright holder]
}

}
\keyword{internal}
