% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrse.R
\name{wrse}
\alias{wrse}
\title{WRSE}
\usage{
wrse(db, L = .Machine$double.xmax)
}
\arguments{
\item{db}{(data.frame) with column \code{X} corresponding to treatment
arms in the first stage (X = 1 for arm A1, X = 2 for arm A2, ... X
= <n> for arm A<n>), \code{R} for respondent or not in the first arm (R
= 0 for not respondent, R = 1 for respondent), \code{Z} for treatment
arms for the second stage (Z = 1 for first arm B1, Z = 2 for second
arm B2), \code{U} for observed time survival (event or censoring time),
\code{delta}event indicator (delta = 0 for censored, delta = 1 for
events).}

\item{L}{(dbl, .Machine$double.xmax) restricted survival time.}
}
\value{
(\link{wrse}) object (inherited from \link{dtr})
}
\description{
WRSE
}
